requirements.txt
av
dataclasses==0.6
dataclasses-json==0.5.2
dvc
joblib
librosa==0.8.1
numpy==1.20
pandas==1.2.4
pyyaml
torch==1.10.0
torchvision==0.11.1
torchaudio==0.10.0
torchtyping==0.1.2
typeguard
typing-extensions==3.7.4.3
pytorchvideo==0.1.2
wandb
opencv-python-headless==4.5.3.56
tensorflow==2.7.0
tqdm
typer
